chainparams: use SeedSpec6's rather than CAddress's for fixed seeds
This negates the need for CAddress here at all
